backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
psycopg2-binary==2.9.6
python-dotenv==1.0.0